[ADD] useInterviewRecord custom hooks for Separation of concerns, redability, reusable
