+++ src/pages/VoiceToTextTestPage/index.js
@@ -1,174 +1,44 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from "react";
+import React, {
+  useCallback,
+  useEffect,
+  useMemo,
+  useRef,
+  useState,
+} from "react";
 import { io } from "socket.io-client";
 
-import { BUTTON_NAME, RECORD_STATE_TYPE } from "../../constants/recordState";
-
-const SpeechRecognition =
-window.SpeechRecognition || window.webkitSpeechRecognition;
+import { BUTTON_NAME } from "../../constants/recordState";
+import useInterviewRecord from "../../hooks/useInterviewRecord";
 
 export default function VoiceToTextTestPage() {
   const socket = useMemo(() => io.connect("http://localhost:5000"), []);
+  const recordBtnElementRef = useRef();
 
-  const [recogText, setRecogText] = useState([]);
+  const [isInterviewee, setIsInterviewee] = useState(false);
+  const [userId, setUserId] = useState("");
 
-  const recordsGlobalsRef = useRef({
+  const {
+    recordStateType,
     recogText,
-    questionText: "",
-    isRecording: false,
-    recognition: null,
-    isInterviewee: false,
-    userId: null,
+    setNextRecordStateType,
+  } = useInterviewRecord({
+    socket,
+    recordBtnElementRef,
+    userId,
+    isInterviewee,
   });
 
-  const recordBtnElement = useRef();
-
-  const [recordBtnState, setRecordBtnState] = useState(
-    RECORD_STATE_TYPE.INTERVIEW_BEFORE
-  );
-
-  const startRecognitionRecord = useCallback(() => {
-    if (!SpeechRecognition) {
-      console.error("Speech recognition not supported 😢 (Use Chrome Browser)");
-      return;
-    }
-
-    if (recordsGlobalsRef.current.isRecording) {
-      return;
-    }
-
-    console.log("recognition start");
-
-    const recognition = new SpeechRecognition();
-
-    recognition.lang = "ko";
-    recognition.continuous = true;
-    recognition.interimResults = true;
-
-    recognition.onstart = () => {
-      setRecogText([]);
-    };
-
-    recognition.onresult = (event) => {
-      const transcript = [...event.results].reduce(
-        (acc, result) => acc + result[0].transcript,
-        ""
-      );
-
-      setRecogText(recogText.concat(transcript));
-    };
-
-    recognition.start();
-
-    recordsGlobalsRef.current.recognition = recognition;
-    recordsGlobalsRef.current.isRecording = true;
-  }, []);
-
-  const stopRecognitionRecord = useCallback(() => {
-    recordsGlobalsRef.current.isRecording = false;
-    recordsGlobalsRef.current.recognition && recordsGlobalsRef.current.recognition.stop();
-    recordsGlobalsRef.current.recognition = null;
-  }, []);
-
-  const setNextRecordBtnState = useCallback(() => {
-    switch (recordBtnState) {
-      case RECORD_STATE_TYPE.INTERVIEW_BEFORE:
-        setRecordBtnState(RECORD_STATE_TYPE.QUESTION_BEFORE);
-        socket.emit("startInterview");
-        break;
-      case RECORD_STATE_TYPE.QUESTION_BEFORE:
-        setRecordBtnState(RECORD_STATE_TYPE.QUESTIONING);
-        startRecognitionRecord();
-        socket.emit("question", { userId: recordsGlobalsRef.current.userId });
-        break;
-      case RECORD_STATE_TYPE.QUESTIONING:
-        setRecordBtnState(RECORD_STATE_TYPE.ANSWER_BEFORE);
-        stopRecognitionRecord();
-        recordsGlobalsRef.current.questionText = recordsGlobalsRef.current.recogText.join(" ");
-        break;
-      case RECORD_STATE_TYPE.ANSWER_BEFORE:
-        setRecordBtnState(RECORD_STATE_TYPE.ANSWERING);
-        socket.emit("requestAnswer");
-        break;
-      case RECORD_STATE_TYPE.ANSWERING:
-        setRecordBtnState(RECORD_STATE_TYPE.SAVING);
-        socket.emit("endAnswer");
-        break;
-      case RECORD_STATE_TYPE.SAVING:
-        setRecordBtnState(RECORD_STATE_TYPE.QUESTION_BEFORE);
-        break;
-      default:
-        setRecordBtnState(RECORD_STATE_TYPE.QUESTION_BEFORE);
-        break;
-    }
-  }, [recordBtnState, socket, startRecognitionRecord, stopRecognitionRecord]);
-
-  useEffect(() => {
-    console.log("set Socket Event");
-
-    socket.on("startInterview", () => {
-      if (!recordsGlobalsRef.current.isInterviewee) {
-        setRecordBtnState(RECORD_STATE_TYPE.QUESTION_BEFORE);
-      }
-      console.log("timer start");
-    });
-
-    socket.on("preventButton", () => {
-      recordBtnElement.current.disabled = true;
-    });
-
-    socket.on("intervieweeStartAnswer", () => {
-      if (recordsGlobalsRef.current.isInterviewee) {
-        console.log("intervieweeStartAnswer, start Record");
-        startRecognitionRecord();
-      }
-    });
-
-    socket.on("intervieweeEndAnswer", () => {
-      if (recordsGlobalsRef.current.isInterviewee) {
-        console.log("intervieweeEndAnswer, stop Record");
-        stopRecognitionRecord();
-        console.log("send Anser", recordsGlobalsRef.current.recogText);
-        socket.emit("sendAnswer", { answer: recordsGlobalsRef.current.recogText });
-      }
-    });
-
-    socket.on("questionerReceiveAnswer", ({ questionerId, answer }) => {
-      if (questionerId === recordsGlobalsRef.current.userId) {
-        console.log("questioner receive Answer uploading...");
-        console.log("question", recordsGlobalsRef.current.questionText);
-        console.log("answer", answer.join(" "));
-        console.log("uploading work...");
-
-        setTimeout(() => {
-          socket.emit("uploadComplete");
-        }, 1500);
-      }
-    });
-
-    socket.on("enableButton", () => {
-      recordBtnElement.current.disabled = false;
-    });
-
-    socket.on("error", ({ message }) => {
-      alert(message);
-    });
-  }, [socket, startRecognitionRecord, stopRecognitionRecord]);
-
-  useEffect(() => {
-    recordsGlobalsRef.current.recogText = recogText;
-  }, [recogText]);
-
   function handleClickRecord() {
-    setNextRecordBtnState();
+    setNextRecordStateType();
   }
 
   const handleKeyDown = useCallback(
     (event) => {
       if (event.key === " " || event.key === "Spacebar") {
-        setNextRecordBtnState();
+        setNextRecordStateType();
       }
     },
-    [setNextRecordBtnState]
+    [setNextRecordStateType]
   );
 
   useEffect(() => {
@@ -187,7 +57,7 @@ export default function VoiceToTextTestPage() {
       },
     });
     const uuid = uuidv4();
-    recordsGlobalsRef.current.userId = uuid;
+    setUserId(uuid);
     console.log("join socket as intervewer", uuid);
     alert("인터뷰어로 참가" + uuid);
   }
@@ -199,7 +69,7 @@ export default function VoiceToTextTestPage() {
         isInterviewee: true,
       },
     });
-    recordsGlobalsRef.current.isInterviewee = true;
+    setIsInterviewee(true);
     console.log("join socket as interviewee");
     alert("면접자로 참가");
   }
@@ -214,9 +84,13 @@ export default function VoiceToTextTestPage() {
       </button>
       <hr />
       <br />
-      {!recordsGlobalsRef.current.isInterviewee && (
-        <button id="button" onClick={handleClickRecord} ref={recordBtnElement}>
-          {BUTTON_NAME[recordBtnState]}
+      {!isInterviewee && (
+        <button
+          id="button"
+          onClick={handleClickRecord}
+          ref={recordBtnElementRef}
+        >
+          {BUTTON_NAME[recordStateType]}
         </button>
       )}
       <h1>녹음 텍스트 : {recogText.join(" ")}</h1>
